+++ rust/main/src/cfg/mod.rs
@@ -14,9 +14,15 @@ pub struct Cfg {
     pub keep_spaces_between_attributes: bool,
     /// Keep all comments.
     pub keep_comments: bool,
-    /// If enabled, CSS in `<style>` tags and `style` attributes are minified.
+    /// Minify CSS in `<style>` tags and `style` attributes using [https://github.com/Mnwa/css-minify](css-minify). By default, the optimisation level is 1 as specified by the CSS minifier, but this can be adjusted by the minify_css_level_* settings.
     pub minify_css: bool,
-    /// If enabled, JavaScript in `<script>` tags are minified using
+    /// Use optimisation level 1 for the CSS minifier. This is currently the default, but may change in the future if higher levels become safe.
+    pub minify_css_level_1: bool,
+    /// Use optimisation level 2 for the CSS minifier. This is mostly safe, but may perform some dangerous optimisations.
+    pub minify_css_level_2: bool,
+    /// Use optimisation level 3 for the CSS minifier. This performs many dangerous optimisations, so ensure any input works with this level.
+    pub minify_css_level_3: bool,
+    /// Minify JavaScript in `<script>` tags using
     /// [minify-js](https://github.com/wilsonzlin/minify-js).
     ///
     /// Only `<script>` tags with a valid or no